One promote command for region selection
Extension publishers can choose to publish to one or more regions not just
two.  The CLI did not enable this behavior to our detriment.  The options
have changed to support --promote one or more time.  Publishers can better
control the list of regions an extension will be published to (or not).