Add missing XML elements
The latest version of azure-extensions-cli is not tolerant of invalidate
manifest documents.  It previously treated the document as a text file
instead of as an XML document.  If the user defined an XML element that
was not recognized by the tool it was fine, and simply passed through.
When the tool started parsing the document it needed to account for
every XML element.  If an element was encountered that it did not
recognize it would ignore it.

The following XML were added.

  * Endpoints
  * LocalResources
  * DisallowMajorVersionUpgrade