No HTTP retries.
The SDK hard codes the number of retries to five, and there is no way to
control this value.  The image APIs generally return failures because
they are "busy."  The SDK ends up sending five requests in quick
succession, which just hammers the server unnecessarily. Because there
is not way to control the value and because it is more likely to fail I
set the number of retries to zero.